+++ src/index.js
@@ -5,12 +5,14 @@ import { Provider } from 'react-redux';
 import { store } from './store'
 import App from './components/App';
 import './index.scss';
+// import Home from './components/Home';
 
 const root = ReactDOM.createRoot(document.getElementById('root'));
 root.render(
   <BrowserRouter>
     <Provider store={store}>
         <App />
+        {/* <Home/> */}
     </Provider>
   </BrowserRouter>
 );
