+++ src/pages/HomePage/HomePage.jsx
@@ -3,9 +3,7 @@ import Home from '../../components/Home';
 
 function HomePage() {
   return (
-    <section className='homepage page'>
-      <h1 className='homepage__title'>HomePage</h1>
-
+    <section>
       <Home />
     </section>
   );
