--- src/pages/HomePage/HomePage.scss
@@ -1,8 +1,0 @@
-.homepage {
-  padding: 10vh 15vw 10vh 15vw;
-  
-  &__title {
-    font-size: 40px;
-    margin-bottom: 5vh;
-  }
-}
